Add files via upload
Two commands:
(1) SCRoF with data or with correlation matrix and sample size as argument(s) and a name for output structure as value. For instance AS=SCRoF(R,N);
(2) SCRoFreport with name of structure as argument to view a list of scenarios in decreasing order of fit probability, e.g. SCRoFreport(AS), or with structure name and rank of a scenario, e.g. SCRoFreport(AS,1), to print the associated solution and paint the dendrogram of pairwse signal cancellation. 
The method is explained and illustrated in a 40 minute video at https://achim.uqam.ca/SCRoF.mp4